capacity to skip external data-fetching step
stringify_model_view() to write out the m object as JSON

modifications to read_model to look for this JSON data in the document
object and load it up if available, skipping all those d3.csv and
d3.json calls, thus bypassing the need for a webserver altogether.

insert_model.py: simple python script to interleave the json data from a
file.

Some explanation in the README.