major reorganization
Access to the model is now via methods on the model object. The data
itself is hidden in a private object known only to the methods.

The datafiles are now loaded incrementally, when the views that need
them are accessed. Probably not the best design, since the result
is a big delay on the first click-through to a topic. The actual loading
is via load_data(), which is called by the *_view() functions.

The model has setter methods that accept the string data from the files.

Some streamlining in the data formats: the metadata is now expected to
have only the columns that are actually used, and the doc-topics matrix
is expected in compressed-column format, which is unrolled when the file
is received. Not sure if this is the best strategy either.