move all rankers into model object
in an attempt to gain back some speed, move the functions for getting top n
words, documents, topics into the model and make use of the sparse form
of dt where appropriate.

also did some clean-up to ensure that ties for last are respected in
each case: this is done via utils.shorten(), which lives in a new source
file, utils.js.