working zip/unzip
Win back 50% of space using zipped data, which is unzipped using JSZip.

load_data() now looks at the requested filename and unzips if it is a
.zip. The data is requested as an arraybuffer using d3.xhr.

dfb.js now loaded before other js files to provide global topic-browser
preferences: it creates a dfb global variable. In particular, it points to
locations of the data files (and whether they are zipped or not).