new overviews using word clouds: scaled, grid
The new model overview offers a choice between the old list or one of
two ways of arranging little word clouds for each topic. One uses
topic_scaled.csv, a file of coordinates. The other is simply a grid.

The major changes are in index.js. model_view is reorganized and split
up into several functions. index.html is also changed around in order to
accommodate a nav row of pills for the model subviews.