Add a random_placement flag to d3.layout.cloud
Davies's code choose a random starting point for each word, which it then tries to offset to avoid collisions. This works well for lots of words in a big space, which gradually get packed in nicely. For a few words in a small space, I am getting somewhat nicer results by just starting all the words at the center of the cloud and letting the algorithm move them from there. So that is what cloud.random_placement(false) allows.