revise model.valid_year
now we take the easy way and just ask for the topic time series and
check whether it has an entry for the given year. This takes advantage
of the cacheing. Shouldn't be too painful on performance, since normally
we're only validating years when we're about to display the time series
anyway.