tooltip revisions
tooltip is now a div added by the javascript to the body (so that it
doesn't get squashed up the edge of the topic_plot container).

show just the year; changed tooltip.text() to expect the data element
instead of just the year--but decided against including the percentage
in the tooltip text after all.

also, some tooltip restyling.