model_view_yearly
a first working yearly model view, a table of top topics for each year.

In addition to the code for model_view_yearly() and style elements for
div#model_view_yearly, this required a little rewriting of the
model_view dispatch itself.

Added a reverse-sort button for the model_view_yearly and then realized
it might as well apply to the model_view_list as well.

Currently topic proportions are hackishly displayed as a colored bar
whose height is proportional to their weight in the year (with
hard-coded parameters). The color is one of 20, determined from taking
t % 20. It ain't pretty.