first working version of topic-word viz
tried but failed to do a version using a regular old table. This does
the table using svg instead. The tilted words are not SO easy to read
but it does allow me to fill the page with information.

FIXME: the transitions in both x and y produce a lot of clutter and need
smoothing out.