experimental: topic_docs moved into a Worker
which requires some refactoring. the dt object methods get moved off into a
separate file. For this first pass, I have left duplicate copies of the
doc-topics matrix, one in the model (main thread) and one for the
worker, so I wouldn't have to make every single dt access call out to
the worker. But m.topic_docs() now takes a callback that receives
the result of the calculation. This is used in topic_view_docs() to
render the table of top docs for the topic asynchronously.

The worker in js/model_worker.js responds to messages sent by the model
object in the main thread. So far it only knows "topic_docs," but
eventually it should hold all the model data and handle all the lengthy
computations. The model stores a d3.map of callbacks on its interface to
the worker (this is needed so parallel requests don't cross).

It's not possible to use the d3 array utility routines in the worker,
because there is no "document" object in the worker's thread and d3
requires one. So I've copied those routines into utils, which can be
in the worker's thread.

The topic_view should give a visual indication that a calculation is in
progress. index.js has a view_calculating function but it doesn't do
anything yet.