progress in refactoring: view.topic
Putting the model in another thread makes it necessary to work out more
carefully what data the views actually need. This leads pretty naturally
to an MVC-style architecture, so I'm going to go ahead and separate out
the view code completely. There is now a view object which accepts
messages from index.js to draw views. So far the only messages it knows
about are the parts of the topic view (view.topic, view.topic.words,
etc.).

Also moved doc_topics off into the worker, with the requisite
modifications to the doc_view (but not yet factoring that out
completely).

Also: topic_words now returns a list of objects rather than just a list
of words that have to be passed back to the model, in preparation for
further refactoring.

In this intermediate stage, there are two copies of the document topic
matrix: one is held in the Worker thread, and one is still held in the
main thread.

Finally, this gave me an opportunity to reconsider the render_updown
architecture and buttons, which produce more visual interference than
they're worth. They are partly but not entirely removed in this commit:

- topic_view: words and docs use the settings held in the VIS object
- doc_view: just show all the doc_topics.