remove duplicate dt object in model
all doc-topics work is now done in the worker thread

slight modification to loading process and flags: m.set_dt now takes a
callback to be executed when the dt is ready. This now has the task of
refreshing the view when the data is in.

also m.dt() checks are replaced with m.has_dt().