spinning gear icon for loading/calculating
the distinction is not important to the user. Also the browser moves
quickly enough that one doesn't see the graphic for more than a split
second in local testing. Animated gif from http://www.ajaxload.info/.

Unfortunately the navbar got too crowded with this. Changed "Problems?"
to "Help!" We will take it out altogether when we are done with the
testing round.