working bib view options with URLs
simplified bib view (and somewhat restyled). navbar is just two popup
menus now, one for sorting and one for asc/desc. All of them use links
so that each bib sort is linkable.

instead of the collapsing panels, an affixed block of links jumps you
around the listing. The affix seems to flicker a bit when you change
sorts. This might be fixed in a later bootstrap version...or is it my
fault?