topic hiding
a not entirely elegant implementation. Topic hiding is governed by two
parameters in VIS (supplied from info.json): show_hidden_topics
(boolean) and hidden_topics (array with topic numbers INDEXED FROM 1 to
hide).

Within the browser the mechanics are, embarrassingly, heterogeneous,
depending on the view. In general the controller functions in main.js
have most of the responsibility for sending the views a filtered list of
topics to display. But for some of the views---model_list and doc---it
is easier to just use a CSS class ("hidden_topic") to hide table rows,
and accordingly the setting trigger manipulates this elements with this
CSS class as well as toggling the flag on VIS.

The plot model overviews transition more or less gracefully when topics
are hidden and shown. The same cannot be said of the yearly overview,
which is simply redrawn (and with different colors).

I have added some descriptive help text about hiding topics on the
about page ("depth" section).