view annotations
Slightly cleaned up view_refresh to call all of the *_view functions in
a uniform way. Now the "switch" statement looks pretty silly, but that's
okay. The job of parsing the parameters in the #/x/y/z URL from strings
into whatever is now that of the *_view functions, so topic_view now
gets called with the 1-indexed topic and has to know to convert it.

The annotations mechanism is very simple. view_refresh(m, "#/x/y/z")
sets anything with a CSS class of "view_x[_y[_z]]" to visible. A global,
VIS.annotes, stores the last set of such classes.

Notice that this fails in cases where the final view is something other
than that specified by "#/x/y/z" (if this is ill-formed or
underspecified). That's a bug to be fixed.