fluid topic view plot
the topic view time series plot now takes its dimensions from the
surrounding topic_plot div. On resize, it transitions smoothly. I
experimented with an animated transition as you go from one topic to
another, but this is quite distracting when you are coming
back from e.g. the word view and suddenly the graph is the old topic,
but transitioning.

This entailed some adjustments to view parameters. VIS.topic_view.w
is now the minimum width of the plot.