fluid word view
The word view responds to the size of its container. VIS.word_view.w is
a minimum width. Because this has to be fairly large, we want to degrade
a little more gracefully than elsewhere. So below the minimum width, the
coordinate system for the word-topics visualization doesn't change, but
it is clipped to the edge of the window. I suppose this is a little mean
to small devices, but the alternative is a mega-squashed visualization.