topic top-words refreshed after setting change
this was a regression, I think. The overlapping "view.dirty" and
"view.updating" mechanisms are both still around and need rationalizing,
but not today.

Hidden-topics checkbox is also itself hidden if there ARE no hideable
topics.

Slightly changed the way the settings form elements are selected (to
make the checkbox-hiding work).