formalize controller as browser()
Most of the functions in main.js are now member functions of the object
returned by browser(). As a result, all the *_view functions no longer
take the model as a parameter; they simply refer to the my.m private member. This is cosmetic but marginally less silly as a way to organize this.

The outward-facing change is that instead of expecting the whole thing
to start with a call to main(), we now prescribe

browser()
    .metadata(dfr_metadata())
    .load();

That is, the end user explicitly takes responsibility for specifying
what kind of metadata is being used.

main() was a dumb name for the loading routine: this is JavaScript, and the
real event loop is inexplicit. When load() returns, the browser
continues to "run."

Finally, the question of multiple authors. Instead of converting
multiple authors in advance when we load metadata, we do it whenever we
cite. It looked awkward to pass "VIS.bib.author_delimiter" to
dfr_metadata on construction or, worse, on load in the middle of
browser().load().