refactoring of bib
A first attempt to consolidate the bibliographic wrangling, so that all
assumptions about the metadata and how to display it are grouped
together. First of all, none of the `view`s should access `bib`
functions: those should be passed from the controller. Second of all, we
should separate generic bib-sorting logic from the specific logic of
sorting DfR journals. This latter aim is realized by having creating a
generic template bib() and then a more specialized bib.dfr(), on the
model of metadata() and metadata.dfr() (N.B. the change from underscores
to dots).

So now the `browser` controller also needs to initialize bib. Some new
logic: some bib configuration variables come from info.json, so these
have to be set once that file is loaded. On the plus side, we can
eliminate the multiply redundant specification of sorting options:
instead of having them once in HTML, once in VIS.bib, and once bib.sort,
we provide one description as bib.sorting(). We still have to write the
key functions. view.bib.dropdown() sets up the sorting option dropdown
from this information.

This clarified that setup_vis() was a bad name for a method whose main
function is to, as it is now called, setup_listeners(). The properties
of VIS are set from info.json in browser().load(), now, not in this
setup routine.

Since this was a big rewrite anyway I did some minor linting and whitespace
cleanup too here and there.