replace global set_view with dfb.set_view
The somewhat inelegant strategy adopted here is for the dfb()
constructor to explicitly give a reference to itself to view. Then
view.dfb().set_view makes this reference available. It's still global
but it makes clear that we're passing a message up to the controller.

view_refresh(window.location.hash) was redundant and has been replaced
by refresh() (window.location.hash is now assumed).