hide vocab of hidden topics in words_view
otherwise, the words_view lists words which lead to a word_view with no
topics. This necessitated a modification to model().vocab() so that it
can accept an array of topic numbers to get vocab for. words_view also
has to re-render each time and can't just be a one-off.