topic_conditional, abstracting topic_year
There is no reason for the model to be committed to summing specifically
over years. As long as we have a partition of document IDs we can generate conditional topic weights for each class in the partition. Here this is implemented by giving metadata a conditionals() method. When we model.set_meta, the model's worker then stores the keys which indicate which documents belong to which partition. In the case of topics by year, the keys are just the years derived from the doc.date field (as strings).

Then we can ask for topic weights and topic totals by partition (e.g.,
year) with topic_conditional, conditional_total. We also need to
generalize getting topic's top documents within a given slice of the
partition with topic_docs_conditional. On the worker side, the unconditioned topic_docs now just calls this routine with an undefined partition key.

Only one user-visible change to functionality. Because only the worker
remembers the partitions, I have removed model.valid_year. It would be
perfectly possible to have model.valid_condition implemented as a
routine taking a callback. But the only place valid_year was used was in
dfb.topic_view, and it was pretty superfluous. I have removed it. The
consequence is that a user who deliberately visits #/topic/24/314159 or
something like that will no longer get the list of top documents.
Instead, they will get a display telling them there are no top documents
for 314159. Which I think is a pretty clear way to handle the edge case.