factor out year-specific stuff from stacked_series
as with view.topic, so too with view.model.yearly: the underlying
calculation should be generic for any condition.
view.model.stacked_series now has this form (and requires, explicitly,
the levels of the conditional variable both as keys to hashes and as
numeric values for the x axis). view.model.yearly is still
particularized as the name implies, so this is not yet fully factored.

also, letting the settings clear VIS.model.yearly.data was...not good.
This uses the view.dirty() mechanism. view.model.yearly now does the
caching of the stack-series calculation, just using a member variable
view.model.yearly.data. Honestly, the view.dirty, view.updating, and
VIS.ready mechanisms all overlap in a big mess.