allow conditional specification in info.json
Instead of the metadata constructor immediately fixing the conditionals,
we can just add them at will up until the moment the model wants to
enumerate them and store document-key values. This can now be done with
the metadata.condition() method, which is invoked once info.json is
loaded. This invocation makes use of a new property on VIS,
VIS.condition, which in turn is keyed to the metadata.key methods, newly
abstracted out and moved from metadata/dfr.js to metadata.js:
metadata.key.category and metadata.key.time. The VIS.condition.key picks
between these two and VIS.condition.spec is passed in as parameter. Thus
we get conditioning on months with

VIS.condition = { key: "time", spec: "%Y-%m" };

or in info.json of course. This replaces the metadata.dfr({ time_key:
"%Y-%m"}) I tried out in the last commit. We still pass objects and not
constructor functions in index.html, but no more parameters (too messy).

This way does add more specific setup stuff to the dfb.load method. Too much happens there, but it'll do for now.

By way of compensatory abstraction, we give dfb a member variable
my.conditional which takes the place of "time" (and is just set to
VIS.condition.key).

API tweak: instead of model.conditionals() we have
model.meta_condition() to give the controller access to the key-inverter
function to pass down to the views.