arbitrary time intervals
tried a few versions of this before settling on a dumb brute-force
solution that allows us to bin the time variable flexibly (in units of 2
years or 3 months or 240 days, etc), at the cost of storing an array of
bin boundaries. Metadata.key.time also has some default choices for key
formats. It can also just numerically bin dates as millisecond values
instead, but that may have wacky results.

Also, the time-bin specification determines what the "step" should be in
the bar charts in #/topic and #/model/list, so we don't have to double
specify that. Instead, the view functions copy the information over from
VIS.condition.spec. A bit tangled.

API change: In order to set up the binning function, we have to know
the minimum date. So the key function has to know about the document
metadata. In practice this means that metadata.condition(k, f, spec)
now takes a third parameter, spec, and stores its key object as the
result of f(spec, docs). Thus a metadata.key generator knows about all
the metadata at construction time. But this also means we have to delay
calling metadata.condition until the metadata is actually loaded. This
rearranges and embellishes the specific program logic in dfb.load(),
which will now do just this before invoking model.set_meta.