change model list table header acc. condition type
Displaying a year range was nice, but for arbitrary time ranges the
dates don't look so nice. Could tweak that but it's a bit silly
to fuss over it. Instead, we just pass in the condition type to
model.view.list and then set the column header accordingly ("over time"
or "by {condition}"). This could be refined.