dfb view functions in a hash
I got tired of the switch statement in `refresh`. dfb() owns a d3.map,
my.views, with keys that are the names of views and functions that are
invoked by `refresh` as before. Experimentally, made the *_view
functions into anonymous functions that are stored in my.views.

The default_view logic has to change a little too, and we now store the
default_view in `my` once we've loaded `VIS`. We try to validate it too.