"year(ly)" -> condition(al) everywhere
escalating the abstraction here by using "conditional" everywhere (and
"condition" for year). Pretty straightforward terminology game here, but
not that we have to change a UI label: "Yearly" in the model nav is
now called "Stacked." Note that we have to change some IDs and CSS
classes in index.html and index.css as well as in the code.

Also there is a serious "API" change in that /model/yearly is called
/model/conditional now. TODO: allow an alias.

Also: utils.clone is a new function for cloning objects, used to copy
model specs rather than munging VIS.topic_view or whatever. So far only
used in one place. TODO: use more widely.