yet another metadata/bib construction reshuffle
if we delay creating dfb's metadata and bib objects until we've loaded
info.json, we can use VIS to decide which class to instantiate. So let's
do that even though it adds more complicated logic to load().
VIS.metadata.type specifies "dfr" or "base"; VIS.metadata.spec gives any
parameters (added examples to src/VIS.js).

bib.options was a bad idea. Now we just pass in options on construction
(in dfb().load). The options are VIS.bib.

This construction step is skipped if custom objects are passed in. But
now we are going to eliminate that annoying step by default, so the
index.html call is back to dfb().load() with no parameters.