VIS.condition needs a field parameter
naming the metadata column to be used for conditionining.
metadata.key.time should not assume it's called "d.date" but should
refer to this field name.

Calling it "date" by default to get the thing to keep working, but n.b.
the column name is set in metadata.dfr().from_string and not (as might
be supposed) in any csv headers. Not altogether user-friendly.