conditioning on ordinal variable
Filling in the details on "ordinal" (categorical) metadata conditions.
On the topic view (and the mini-barplots in the list view), we have to
do a bunch of fiddly adjustments to the axis and x scale.

In the stacked model view, the streamgraph is a bad idea for categorical
values, so we do stacked bars (ordered by topic number). d3 doesn't give
us an easy way to zoom an ordinal scale, so zooming and panning only
works along y here.

metadata key functions introduce a third sub-function, `display`.
We might want to use one string key to label categories/intervals
internally and another when we label categories for the user in
the topic view. metadata.key.time.display spells out the interval
instead of just giving the left endpoint. metadata.key.ordinal
replaces spaces with underscores in keys, but the .display variant
puts them back (this way we get `#/topic/Modern_Philology` instead of
`#/topic/Modern%20Philology`)