get condition by continuous variable working
Needed to be more careful about metadata.key.continuous: when we bin and
stringify continuous numbers, we'd better try to preserve ordering. So
this has some heuristics for guessing how many leading zeroes and
decimal places we need for that. Haven't tested it with anything but a
set of positive integers.