metadata.key.range (used in #/model/conditional)
the key converter for a conditioning covariate needs to say what its
full range of possible outputs is in order to deal with the possibility
that some intermediate bins of a continuous covariate are empty for all documents. This is not a
problem on the individual topic plots, but on the streamgraph it yields
an incorrect display, where the topic proportions are shown as constant
across the empty bin. So metadata.key.range is a superset of the keys of
model().conditional_totals, which we used to use for this purpose.

With my test pagelength data, the streamgraph is very ugly across the
empty bins, but I *think* this is not an error in the way the display is
rendered, just an aesthetic failure.