topic view time axis ticks: sane default
Replicate the two options offered by d3.time.scale.ticks: if
VIS.topic_view.time.ticks is a number, treat it as a number of ticks;
if it is an object with `n` and `unit` fields, treat it as an interval.
We'll use 10 ticks as the default instead of 10-year intervals, which is
only a good choice if you have about a century's worth of stuff.