dfb().setup_views
Begin to factor out view setup. We mark all views as dirty to start.
This is still not perfectly systematic, but it will do for the
moment. We use the dirty mechanism on view.about.

This also introduces the frame view, which is where we now set the model
title element (from with setup_views instead of within load()).