cur_view processing only in refresh(); bugfix
doesn't need to be global on VIS, can just be a my.cur_view.

The bugfix is to the heuristic for deciding on whether we're `updating`:
Can't use regex match because `word` matches `words` and we get both
views visible when shifting from #/words to #/word/whatevs. Let's be
less silly and just store the first component of #/view/p1/p2.