check for normalized ("proper") doc-topics
The underlying code works fine if the doc-topics matrix is normalized
rather than raw sampling weights. But we should change some of the
interface labeling, since it's misleading to display the total "tokens"
in a document as 1 (or 0.9999). Also the #/topic bar charts are not
"conditional prob of words in topic" but of docs. And the stacked
display changes meaning too.

Normalization is detected when we set_dt. Since we're already
precalculating row sums, we just check they're all close to 1. (How
close: PROPER_TOLERANCE is hard-coded in worker.js).

The result of the check passes back up to the controller which passes
it on to the views as needed (#/doc, #/topic, which then choose not
to create a "tokens" column for documents). The controller also
explicitly hides/shows everything with css classes proper/not-proper
as appropriate. This lets us make the remaining interface changes by
writing HTML in index.html.