refine svg-append behavior
caught a bug in model/list when sorting a list with hidden topics: the
data selection needed a key function. While fixing the bug, became
dissastisfied with the un-data-join-like approach to adding the
sparkplots.

Fixing that required changing everything that used view.plot_svg (and
its helper view.append_svg). These are now replaced with two functions,
view.append_plot and view.setup_plot. The former is meant to be called
on an entering selection of svg's and the latter on an enter + update
selection. All the margin-convention work is done by setup_plot, which
never appends elements. Single-svg views just bind `[1]` to an `svg`
selection so that we append the svg once and then modify it later.

This should help with model-switching.