get annotations working (as view method)
view.update_annotations() now does this job, not the controller. The
controller should only call it. This gets rid of the stupid VIS.annotes
array which we don't really need. Store the state on CSS classes!

Changed the mechanism a little. Annotations should have the two classes
`annote` and `view_param_param` (e.g. topic_5). This means the user
doesn't have to remember to make annotations hidden to start out.