zeroth draft: multiple models
many changes, mostly in dfb(), to make it possible to switch between
models in a single browser. The multiple-model functions are pure draft
and NOT TESTED here. The only testing I've done is some basic checking
that browsing a single-model with a single set of data files still
works as it should.

A key overall change is that all view-setting and specification of
the form of links to other views is done by a controller method.
Instead of a view generating an "a href="#/topic/2" we want "a
href=#/currentmodel/topic/2" where currentmodel is filled in with the
current model ID. This is done by specifying views as collections of
properties { model: ..., type: ..., param: [...] } which are passed to
dfb().set_view() or dfb().view_link(). These methods are now called by
all of the views.

dfb().parse_hash(h) converts a window.location.hash value h into an
object like this. It is invoked by dfb().refresh() (and in parsing the
default view preference). parse_hash is actually liberal and will turn
"#/topic/2" into "#/curmodel/topic/2" which means the hard-coded view
links in index.html don't need to be changed.

If there is only the one model, we keep the previous style of urls
("#/topic/2").

Data loading is factored into more functions. The main dfb().load()
decides whether we are doing a multiple-model or a single-model browser.
Then it hands off loading to load_model, a method which is meant to
be called each time we switch models. load_model has to invalidate
any caches of information specific to the model, create a new model
object if necessary, and then invoke the now factored-out individual
file loading functions load_info, load_dt, load_tw, load_meta, and
load_topic_scaled. Each of these functions checks whether it actually
has to retrieve a file or whether the necessary information is already
available on the current model object, then does the necessary setup. In
the single-model case, load_info gets passed the already-loaded settings
from info.json.

The whole collection of models is stored in dfb as my.ms, a d3.map keyed
to id values. my.models is the array of model descriptors.

TODO still: each model is a new object with new data. There is no way to
share metadata or model data between models. Shouldn't be too hard to
add, but let's get the thing rolling first.

Managing settings is more complex in this regime, and I have changed
the logic. We have overall app settings and model-specific settings.
In the file structure, these correspond to data/info.json and
data/xyz/info.json, data/zyx/info.json, etc. The settings are formed by
merging data/info.json with data/xyz/info.json. The results are still
stored in the stupid VIS global but at least xyz can't overwrite zyx.
But this means we have to keep separate track of settings which are
specified in info.json but should persist between models. There aren't
that many of these, so far now I have the kludgy solution of listing
them explicitly as my.settings and passing them around where needed.

Thus, now, the VIS global can only change when a model is loaded. No
temporary information or user-mutable settings should be stored on VIS
now.