labels must remain on top of layers in model/cond
this was getting messed up if we changed data sets and drew new layers
on top of the old labels. Instead, we put all layers in a group element
that is drawn before the labels. We can add/subtract from that without
problems.

TODO make transitions in this view less ugly