fiddling model/cond transitions
trying to get things to work a little better. The render_marks and
render_labels functions are now callable on either selections or
transitions. Zoom + transition is still tricky; if the timing is wrong
(tx duration is too small ???), launching a transition at every zoom
event causes things to get "stuck." So instead we just won't transition
on zoom.

Annoying initial delay is annoying. Checking for initial conditions is
a bother when we can't take the size of an enter selection. Guess what?
That was fixed in d3 3.4.5. Time to update.