data-sharing for multiple models
Don't double-load and process when we don't have to. dfb now has two
more data-storage maps, my.shared.models and my.shared.meta, which are
keyed to filenames. (models are keyed to the concat of tw, dt,
topic_scaled). Then when we switch models we check whether we've already
loaded the appropriate stuff for another model.

This necessitated a little tweaking in the interface for the model object wrt metadata and conditions. In particular model().meta() is redefined to get/set the model object itself, not act as a document metadata accessor. That is now model().meta().doc(). Then model().meta_condition() can be a setter as well as a getter and will appropriately invoke doc_category() to store document categories for future tallying. When we do this we need to take a callback so that we have a chance to refresh a topic/conditional plot etc AFTER doing doc_category(), in the case where we switch models but operate on the same metadata (in which case we have model().ready("meta") is true). model().ready("meta_XXX") now gives a meaningful value for condition names XXX.)