dfb().topic_hidden as a method
abstracting away the need to check that my.cache.topic_hidden exists
before indexing into it, and the need to combine this check with
examining the hide/show topic settings. This mechanism is still pretty
messy but this is a bit better. This also catches a tricky race
condition where we swap "models" but keep the same model object. In that
case the data will be ready and a view can attempt to render before the
info.json request and the subsequent loading is finished---including
resetting the topic_hidden array. Parallel programming sux.